requirements.txt
Django==2.1.4
channels==2.1.6
channels-redis==2.3.2
asgi-redis==1.0.0
djangorestframework==3.8.2